pyproject.toml
[project]
name = "TREEEX-WBSP"
version = "0.1.0"
description = "WhatsApp Business Solution Provider platform - A robust API for managing WhatsApp Business messaging, campaigns, and webhooks"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TREEEX Team"}
]
keywords = ["whatsapp", "business", "api", "messaging", "fastapi"]
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "azure-storage-blob>=12.19.0",
    "fastapi[standard]>=0.123.2",
    "httpx>=0.27.0",
    "ngrok>=1.4.0",
    "openpyxl>=3.1.2",
    "pre-commit>=4.5.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.18",
    "redis[asyncio]>=7.1.0",
    "sentry-sdk[fastapi]>=2.46.0",
    "sqlalchemy[all]>=2.0.44",
    "sqlmodel>=0.0.27",
    "supabase>=2.25.0",
    "werkzeug>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
